Containerize R Markdown Documents

Containerize R Markdown documents. This function generates Dockerfile based on the liftr metadata in the RMD document.

lift(input = NULL, output_dir = NULL)

Arguments

input
Input (R Markdown) file.

output_dir
Directory to output Dockerfile. If not provided, will be the same directory as input.

Value

Dockerfile.

Details

After running lift, run render_docker on the document to render the Dockerized R Markdown document using Docker containers. See vignette('liftr-intro') for details about the extended YAML front-matter metadata format used by liftr.

Examples

# copy example file
dir_example = paste0(tempdir(), '/liftr-minimal/')
dir.create(dir_example)
file.copy(system.file("examples/liftr-minimal.Rmd", package = "liftr"), dir_example)

#> [1] TRUE

# containerization
input = paste0(dir_example, "liftr-minimal.Rmd")
lift(input)

# view generated Dockerfile
readLines(paste0(dir_example, "Dockerfile"))

#>  [1] "FROM rocker/r-base:latest"                                                                                                                                        
#>  [2] ""                                                                                                                                                                 
#>  [3] "MAINTAINER Nan Xiao <me@nanx.me>"                                                                                                                                 
#>  [4] ""                                                                                                                                                                 
#>  [5] "# System dependencies for required R packages"                                                                                                                    
#>  [6] "RUN  rm -f /var/lib/dpkg/available \\"                                                                                                                            
#>  [7] "  && rm -rf  /var/cache/apt/* \\"                                                                                                                                 
#>  [8] "  && apt-get update -qq \\"                                                                                                                                       
#>  [9] "  && apt-get install -y --no-install-recommends \\"                                                                                                               
#> [10] "    ca-certificates \\"                                                                                                                                           
#> [11] "    libssl-dev \\"                                                                                                                                                
#> [12] "    libcurl4-openssl-dev \\"                                                                                                                                      
#> [13] "    libxml2-dev \\"                                                                                                                                               
#> [14] "    git"                                                                                                                                                          
#> [15] ""                                                                                                                                                                 
#> [16] ""                                                                                                                                                                 
#> [17] ""                                                                                                                                                                 
#> [18] ""                                                                                                                                                                 
#> [19] ""                                                                                                                                                                 
#> [20] "# Pre-compiled pandoc required by rmarkdown"                                                                                                                      
#> [21] "# Version from: https://github.com/metacran/r-builder/blob/master/pkg-build.sh"                                                                                   
#> [22] "# Scripts from: https://github.com/jangorecki/dockerfiles/blob/master/r-pkg/Dockerfile"                                                                           
#> [23] "RUN PANDOC_VER=\"1.17.2\" \\"                                                                                                                                     
#> [24] "  && PANDOC_DIR=\"/opt/pandoc\" \\"                                                                                                                               
#> [25] "  && PANDOC_URL=\"https://s3.amazonaws.com/rstudio-buildtools/pandoc-${PANDOC_VER}.zip\" \\"                                                                      
#> [26] "  && mkdir -p \"${PANDOC_DIR}\" \\"                                                                                                                               
#> [27] "  && wget --no-check-certificate -O /tmp/pandoc-${PANDOC_VER}.zip ${PANDOC_URL} \\"                                                                               
#> [28] "  && unzip -j /tmp/pandoc-${PANDOC_VER}.zip \"pandoc-${PANDOC_VER}/linux/debian/x86_64/pandoc\" -d \"${PANDOC_DIR}\" \\"                                          
#> [29] "  && chmod +x \"${PANDOC_DIR}/pandoc\" \\"                                                                                                                        
#> [30] "  && ln -s \"${PANDOC_DIR}/pandoc\" /usr/local/bin \\"                                                                                                            
#> [31] "  && unzip -j /tmp/pandoc-${PANDOC_VER}.zip \"pandoc-${PANDOC_VER}/linux/debian/x86_64/pandoc-citeproc\" -d \"${PANDOC_DIR}\" \\"                                 
#> [32] "  && chmod +x \"${PANDOC_DIR}/pandoc-citeproc\" \\"                                                                                                               
#> [33] "  && ln -s \"${PANDOC_DIR}/pandoc-citeproc\" /usr/local/bin \\"                                                                                                   
#> [34] "  && rm -f /tmp/pandoc-${PANDOC_VER}.zip"                                                                                                                         
#> [35] ""                                                                                                                                                                 
#> [36] "RUN Rscript -e \"install.packages(c('devtools','knitr','rmarkdown','shiny','RCurl'), repos = 'https://cran.rstudio.com')\""                                       
#> [37] ""                                                                                                                                                                 
#> [38] "RUN Rscript -e \"source('https://cdn.rawgit.com/road2stat/liftrlib/fab41764ea8b56677d05c70c86225774164b6ca0/install_cran.R');install_cran(c('tufte','kernlab'))\""
#> [39] ""                                                                                                                                                                 
#> [40] ""                                                                                                                                                                 
#> [41] ""                                                                                                                                                                 
#> [42] "RUN mkdir /liftrroot/"                                                                                                                                            
#> [43] "WORKDIR /liftrroot/"